auth_service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
email-validator
python-multipart 
fastapi-mail
gunicorn